+++ src/EventBroker.js
@@ -127,9 +127,6 @@ function EventBroker(brokerOwner, options, onBrokerReturn) {
       case 'wait': {
         return `activity.${eventName}`;
       }
-      case 'error': {
-        return `${pfx}.error`;
-      }
       default: {
         return `${pfx}.${eventName}`;
       }