+++ src/messageHelper.js
@@ -27,7 +27,7 @@ function cloneContent(content, extend) {
   if (outbound) {
     clone.outbound = outbound.map((c) => cloneContent(c));
   }
-  if (sequence) {
+  if (Array.isArray(sequence)) {
     clone.sequence = sequence.map((c) => cloneContent(c));
   }